recostructed admin routes, added recipes SoftDeletes, css changes

+++ routes/web.php
@@ -22,26 +22,26 @@ Route::middleware(['auth', 'role'])->group(function () {
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
-    Route::get('recipes', [RecipeController::class, 'index']);
-    Route::get('recipes/{id}', [RecipeController::class, 'show'])->whereNumber('id');
-    Route::get('recipes/create', [RecipeController::class, 'create']);
-    Route::any('recipes/edit/{id}', [RecipeController::class, 'edit'])->name('recipe.edit');
-    Route::post('recipes/store', [RecipeController::class, 'store']);
-    Route::delete('recipes/delete/{id}', [RecipeController::class, 'delete'])->name('recipe.delete');
+    Route::get('admin/recipes', [RecipeController::class, 'index']);
+    Route::get('admin/recipes/{id}', [RecipeController::class, 'show'])->whereNumber('id');
+    Route::get('admin/recipes/create', [RecipeController::class, 'create']);
+    Route::any('admin/recipes/edit/{id}', [RecipeController::class, 'edit'])->name('recipe.edit');
+    Route::post('admin/recipes/store', [RecipeController::class, 'store']);
+    Route::delete('admin/recipes/delete/{id}', [RecipeController::class, 'delete'])->name('recipe.delete');
 
-    Route::get('categories', [CategoryController::class, 'index']);
-    Route::get('categories/create', [CategoryController::class, 'create']);
-    Route::post('categories/create', [CategoryController::class, 'store']);
-    Route::any('categories/edit/{id}', [CategoryController::class, 'edit'])->name('category.edit');
-    Route::delete('categories/delete/{id}', [CategoryController::class, 'delete'])->name('category.delete');
-    Route::get('categories/{id}', [CategoryController::class, 'show']);
+    Route::get('admin/categories', [CategoryController::class, 'index']);
+    Route::get('admin/categories/create', [CategoryController::class, 'create']);
+    Route::post('admin/categories/create', [CategoryController::class, 'store']);
+    Route::any('admin/categories/edit/{id}', [CategoryController::class, 'edit'])->name('category.edit');
+    Route::delete('admin/categories/delete/{id}', [CategoryController::class, 'delete'])->name('category.delete');
+    Route::get('admin/categories/{id}', [CategoryController::class, 'show']);
 
-    Route::get('ingredients', [IngredientController::class, 'index']);
-    Route::get('ingredients/create', [IngredientController::class, 'create']);
-    Route::post('ingredients/create', [IngredientController::class, 'store']);
-    Route::any('ingredients/edit/{id}', [IngredientController::class, 'edit'])->name('ingredient.edit');
-    Route::delete('ingredients/delete/{id}', [IngredientController::class, 'delete'])->name('ingredient.delete');
-    Route::get('ingredients/{ingredient}', [IngredientController::class, 'show']);
+    Route::get('admin/ingredients', [IngredientController::class, 'index']);
+    Route::get('admin/ingredients/create', [IngredientController::class, 'create']);
+    Route::post('admin/ingredients/create', [IngredientController::class, 'store']);
+    Route::any('admin/ingredients/edit/{id}', [IngredientController::class, 'edit'])->name('ingredient.edit');
+    Route::delete('admin/ingredients/delete/{id}', [IngredientController::class, 'delete'])->name('ingredient.delete');
+    Route::get('admin/ingredients/{ingredient}', [IngredientController::class, 'show']);
 });
 
 require __DIR__ . '/auth.php';
